pyproject.toml
[project]
name = "thesis-os"
version = "0.1.0"
description = "Decision artifacts for fundamental equity PMs"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.30,<1",
    "anthropic>=0.49,<1",
    "httpx>=0.28,<1",
    "cachetools>=5.5,<6",
    "scipy>=1.15,<2",
    "pydantic-settings>=2.7,<3",
    "alembic>=1.14,<2",
    "fpdf2>=2.8,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.25,<1",
    "ruff>=0.9,<1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 99
